Add header setting file encoding and get rid of padding.
Org mode pads tangled code blocks with empty lines by default. This is
stupid in case of the first code block because it results in an empty
line at the beggining of the file.

And what better thing to put at the beginning of the file than file
local variables? So set coding to UTF-8 just in case.